+++ src/core/server/extensions/playerFunctions/inventory.ts
@@ -981,6 +981,7 @@ export default {
     getEquipmentItem,
     getFreeInventorySlot,
     getInventoryItem,
+    getSlotType,
     getToolbarItem,
     handleSwapOrStack,
     hasItem,